- переименование таблицы service - в occupation - создание сервиса работ (occupation) - разрешение конфликтов мержа - удалил Service - замена Integer на Long

+++ src/main/java/com/gb/agile/craft_master/config/JwtProvider.java
@@ -19,9 +19,9 @@ public class JwtProvider {
     private static final String LOGIN_CLAIM = "login";
     private static final String ROLE_CLAIM = "role";
 
-    private static final ThreadLocal<Integer> userId = new ThreadLocal<>();
+    private static final ThreadLocal<Long> userId = new ThreadLocal<>();
 
-    public String generateToken(Integer userId, String login, String role) {
+    public String generateToken(Long userId, String login, String role) {
         Instant expirationTime = Instant.now().plus(60, ChronoUnit.DAYS);
         Date expirationDate = Date.from(expirationTime);
 
@@ -38,7 +38,7 @@ public class JwtProvider {
     public boolean validateToken(String token) {
         try {
             Claims claims = Jwts.parser().setSigningKey(JWT_SECRET).parseClaimsJws(token).getBody();
-            userId.set(claims.get(USERID_CLAIM, Integer.class));
+            userId.set(claims.get(USERID_CLAIM, Long.class));
             return true;
         } catch (Exception e) {
             log.severe("invalid token");
@@ -62,7 +62,7 @@ public class JwtProvider {
 //        tokenRedisRepository.putToken(token, duration);
     }
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return userId.get();
     }
 }
